% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{construct_headers}
\alias{construct_headers}
\title{Construct the headers for the query}
\usage{
construct_headers(api_key)
}
\arguments{
\item{api_key}{The API-key for Hilma.}
}
\value{
An object for the \code{config}-argument of the \code{\link[httr:GET]{httr::GET()}}/\code{\link[httr:POST]{httr::POST()}}-function.
}
\description{
Creates the header needed for a \code{\link[httr:GET]{httr::GET()}}/\code{\link[httr:POST]{httr::POST()}}-request.
}
