% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl.R
\name{fetch_fields}
\alias{fetch_fields}
\title{Fetch api documentation}
\usage{
fetch_fields(api_key)
}
\arguments{
\item{api_key}{Your personal API-key. Can be obtained from
\url{https://hns-hilma-prod-apim.portal.azure-api.net}.}
}
\value{
A tibble with three columns:
\itemize{
\item \code{name}: The name of the field
\item \code{type}: The type of the field
\item \code{searchable}: Whether the field is searchable or not
\item \code{filterable}: Whether the field is filterable or not
}
}
\description{
Fetches the the fields available from the api.
}
