% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl.R
\name{fetch_hilma_data}
\alias{fetch_hilma_data}
\title{Fetch the data from the API}
\usage{
fetch_hilma_data(skip, query, headers, verbose, api_url, n_fetch = 1000L)
}
\arguments{
\item{skip}{How many notices should be skipped from the beginning. The notices
are arranged by \code{id} before the skipping.}

\item{query}{The query parameters. Can be constructed with \link{query-helpers}.}

\item{headers}{The headers of the request. Constructed with \code{\link[=construct_headers]{construct_headers()}}.}

\item{verbose}{Whether to print the progress or not.}

\item{api_url}{Url to the hilma_api.}

\item{n_fetch}{The (maximum) number of notices to fetch. Defaults to \code{1000L}
which is the limit of the API.}
}
\value{
The \code{value}-field of the result in a tibble.
}
\description{
Fetches the data given the query parameters.
}
