% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl.R
\name{fetch_notices}
\alias{fetch_notices}
\title{Fetch notices from the Hilma API}
\usage{
fetch_notices(query, api_key, include_search_score = FALSE, verbose = FALSE)
}
\arguments{
\item{query}{The query parameters as a list that will be encoded as
JSON with \code{\link[httr:POST]{httr::POST()}} using \code{encode = "json"}. Can be constructed with
\link{query-helpers}.}

\item{api_key}{Your personal API-key. Can be obtained from
\url{https://hns-hilma-prod-apim.portal.azure-api.net}.}

\item{include_search_score}{Whether to return the search score (ie. how well
the result matches the query) or not. Defaults to \code{FALSE}.}

\item{verbose}{Whether to print the progress or not. Defaults to \code{FALSE}.}
}
\value{
A tibble that contains the result of the query.
}
\description{
Fetches the notices specified in the \code{\link[=query-helpers]{query}}-argument
from the API and formats the result in a tibble. Errors if the results changes
(ie. someone publishes a notice that matches the \code{query}) during the
execution.
}
