% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl.R
\name{fetch_number_of_results}
\alias{fetch_number_of_results}
\title{Fetch the number of results}
\usage{
fetch_number_of_results(query, headers, api_url)
}
\arguments{
\item{query}{The query parameters. Can be constructed with \link{query-helpers}.}

\item{headers}{The headers of the request. Constructed with \code{\link[=construct_headers]{construct_headers()}}.}

\item{api_url}{Url to the hilma_api.}
}
\value{
An integer, the number of results.
}
\description{
Similar to \code{\link[=fetch_hilma_data]{fetch_hilma_data()}} but returns the number of results rather
than the result itself.
}
