% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{generic_boolean_filter}
\alias{generic_boolean_filter}
\title{A generic filter for boolean value}
\usage{
generic_boolean_filter(variable, is_true)
}
\arguments{
\item{variable}{The name of the variable}

\item{is_true}{Whether to include the notices that match the filter or not.}
}
\value{
A list containing the query that can be passed to \code{\link[=fetch_notices]{fetch_notices()}}.
}
\description{
Searches the \code{variable} for an exact match of the \code{value}.
}
