% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_notice_tibble}
\alias{process_notice_tibble}
\title{Process the data returned from the API}
\usage{
process_notice_tibble(tbl_raw)
}
\arguments{
\item{tbl_raw}{The unprocessed tibble returned from the API.}
}
\value{
A tibble with same dimensions and some times fixed.
}
\description{
Fixes the types of timestamp-strings to \link[=DateTimeClasses]{date-times} and
sets the type of \code{id} to integer.
}
