% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query-helpers}
\alias{query-helpers}
\alias{query_and}
\alias{query_or}
\alias{dps_filter}
\alias{framework_agreement_filter}
\alias{is_national_filter}
\alias{notice_number_filter}
\alias{id_filter}
\alias{procurement_project_filter}
\alias{all_notices_modified_since_filter}
\alias{all_notices_modified_until_filter}
\alias{all_notices_published_since_filter}
\alias{all_notices_published_until_filter}
\alias{organisation_name_search}
\alias{organisation_id_search}
\alias{cpv_search}
\title{Helper functions for constructing the queries}
\usage{
query_and(q1, q2)

query_or(q1, q2)

dps_filter(is_dps = TRUE)

framework_agreement_filter(is_fa = TRUE)

is_national_filter(is_national = TRUE)

notice_number_filter(notice_number)

id_filter(id)

procurement_project_filter(id)

all_notices_modified_since_filter(since = today() \%m-\% days(1))

all_notices_modified_until_filter(until = today())

all_notices_published_since_filter(since = today() \%m-\% days(1))

all_notices_published_until_filter(until = today())

organisation_name_search(name)

organisation_id_search(organisation_id)

cpv_search(cpv)
}
\arguments{
\item{q1}{First query}

\item{q2}{Second query}

\item{is_dps}{Whether to filter notices that are for DPSs or notices that
are not for DPSs.}

\item{is_fa}{Whether to filter notices that are for FAs or notices that
are not for FAs.}

\item{is_national}{Whether to filter notices that are for FAs or notices that
are not for FAs.}

\item{notice_number}{The notice number. E.g. \code{"2020-000001"}.}

\item{id}{The id of the procurement project. E.g. \code{1L}.}

\item{since}{A \link{date}-object. Includes the boundary and defaults to yesterday.}

\item{until}{A \link{date}-object. Includes the boundary and defaults to today.}

\item{name}{The name of the procuring organisation. E.g. \code{"Hansel Oy"}.}

\item{organisation_id}{The national registration id of the procuring organisation.
E.g. \code{"1234567-8"}.}

\item{cpv}{The cpv code to search E.g. \code{"00000000"}.}
}
\value{
A list containing the query that can be passed to \code{\link[=fetch_notices]{fetch_notices()}}.
}
\description{
Functions for constructing the \code{query_params}-argument of the
\code{\link[=fetch_notices]{fetch_notices()}}-function. Functions suffixed with \verb{_filter} and \verb{_search}
use the respective functionalities of the
\href{https://docs.microsoft.com/en-us/azure/search/query-lucene-syntax}{Azure Cognitive Search}.
}
\details{
\code{query_and} and \verb{_or} can be used to combine queries (see example below).
}
\examples{
# A query that returns all notices modified 2020-01-01 or later
# from a certain organisation
# q <- organisation_id_filter("1234567-8") \%>\%
#        query_and(all_notices_modified_since_filter(lubridate::ymd("2020-01-01")))
# fetch_notices(q, api_key)
}
