% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_combine}
\alias{query_combine}
\title{Combine queries}
\usage{
query_combine(q1, q2, pred)
}
\arguments{
\item{q1}{The second query}

\item{q2}{The first query}

\item{pred}{The predicate used for combining the queries.}
}
\value{
A list containing the new query.
}
\description{
Combines two queries with filters and searches using a predicate within the
filter and/or search.
}
