% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{split_into_chunks}
\alias{split_into_chunks}
\title{Split the fetch into multiple chunks}
\usage{
split_into_chunks(n_notices, batch_size = 1000L)
}
\arguments{
\item{n_notices}{The number of notices to be fetched}

\item{batch_size}{The desired batch size, the current default is \code{1000},
which is the limit of the current API.}
}
\value{
An integer vector that contains the number of elements to
skip in each query.
}
\description{
A helper function for splitting the fetch into chunks that the api accepts.
}
